Changed errorRate to a double.

+++ src/com/metrostate/ics460/project2/receiver/DataReceiver.java
@@ -2,5 +2,5 @@ package com.metrostate.ics460.project2.receiver;
 
 public interface DataReceiver {
 
-	byte[] receiveData(String receiverIpAddress, int receiverPort, int windowSize, int errorRate);
+	byte[] receiveData(String receiverIpAddress, int receiverPort, int windowSize, double errorRate);
 }
